Move script handlers out of main.ts
